create markdown details

Create details block for markdown with summary as optional.

details(text, summary = NULL, tooltip = "Click to Expand",
  open = FALSE, lang = "r", output = c("console", "clipr", "edit",
  "character"))

Arguments

text
character, text to put in details block

summary
character, text to put in summary block, Default: NULL

tooltip
character, text for tooltip on the summary, Default: 'Click to Expand'

open
logical, is the details open (TRUE) or closed (FALSE), Default: FALSE

lang
character, language of block (for markdown highlighting) Default: 'r'

output
character, where to output the file console (Default), clipboard or R file editor, Default: c('console','clipr','file.edit')

Value

character

Details

To remove summary or tooltip set them to NULL. If the text is a file path, it will automatically it's lines will be read in internally.

More Details
Here more information can be inserted without overwhelming the user.

To enable this feature add Roxygen: list(markdown = TRUE) to the DESCRIPTION file before rendering the roxygen2

you can use this feature by wrapping documentation with the macros

\foldstart{[SUMMARY TEXT]}

#' DOCUMENTATION

\foldend