Create HTML DOM Details

Create HTLM DOM Details block for Markdown documents with summary as optional.

details(object, summary = NULL, tooltip = "Click to Expand",
  open = FALSE, lang = "r", output = c("console", "clipr", "edit",
  "character"))

Arguments

object
object, object to put in details block

summary
character, text to put in summary block, Default: NULL

tooltip
character, text for tooltip on the summary, Default: 'Click to Expand'

open
logical, is the details open (TRUE) or closed (FALSE), Default: FALSE

lang
character, language of block (for markdown highlighting) Default: 'r'

output
character, where to output the file console (Default), clipboard or R file editor, Default: c('console','clipr','file.edit','character')

Value

character

Details

To remove summary or tooltip set them to NULL. If the object is a file path, it will automatically it's lines will be read in internally.

If lang is NULL then the output will not be wrapped in a code block and will display the raw output (useful for HTML)

Examples

#basic
 details::details('test')

#> <details closed>
#> 
#> ```r
#> 
#> test
#> 
#> ```
#> 
#> </details>
#> <br>

#sessionInfo
 details::details(sessionInfo(), summary = 'sessionInfo')

#> <details closed>
#> <summary> <span title='Click to Expand'> sessionInfo </span> </summary>
#> 
#> ```r
#> 
#> R version 3.6.1 (2017-01-27)
#> Platform: x86_64-pc-linux-gnu (64-bit)
#> Running under: Ubuntu 14.04.5 LTS
#> 
#> Matrix products: default
#> BLAS:   /home/travis/R-bin/lib/R/lib/libRblas.so
#> LAPACK: /home/travis/R-bin/lib/R/lib/libRlapack.so
#> 
#> locale:
#>  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C              
#>  [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8    
#>  [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
#>  [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                 
#>  [9] LC_ADDRESS=C               LC_TELEPHONE=C            
#> [11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       
#> 
#> attached base packages:
#> [1] stats     graphics  grDevices utils     datasets  methods   base     
#> 
#> other attached packages:
#> [1] details_0.1.1  testthat_2.2.1
#> 
#> loaded via a namespace (and not attached):
#>  [1] Rcpp_1.0.2         pillar_1.4.2       compiler_3.6.1     remotes_2.1.0     
#>  [5] prettyunits_1.0.2  tools_3.6.1        highlight_0.4.7.2  pkgload_1.0.2     
#>  [9] pkgbuild_1.0.5     digest_0.6.21      memoise_1.1.0      evaluate_0.14     
#> [13] tibble_2.1.3       pkgconfig_2.0.3    png_0.1-7          rlang_0.4.0       
#> [17] cli_1.1.0          rstudioapi_0.10    commonmark_1.7     curl_4.2          
#> [21] yaml_2.2.0         pkgdown_1.1.0.9000 xfun_0.10          withr_2.1.2       
#> [25] httr_1.4.1         stringr_1.4.0      roxygen2_6.1.1     xml2_1.2.2        
#> [29] knitr_1.25         devtools_2.1.0     desc_1.2.0         fs_1.3.1          
#> [33] rprojroot_1.3-2    glue_1.3.1         R6_2.4.0           processx_3.4.1    
#> [37] rmarkdown_1.16     sessioninfo_1.1.1  purrr_0.3.2        callr_3.3.1       
#> [41] clipr_0.7.0        magrittr_1.5       rematch2_2.1.0     whisker_0.4       
#> [45] ps_1.3.0           backports_1.1.4    clisymbols_1.2.0   htmltools_0.4.0   
#> [49] MASS_7.3-51.4      usethis_1.5.1      assertthat_0.2.1   stringi_1.4.3     
#> [53] crayon_1.3.4      
#> 
#> ```
#> 
#> </details>
#> <br>

#data.frame
 details::details(head(mtcars))

#> <details closed>
#> 
#> ```r
#> 
#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1
#> 
#> ```
#> 
#> </details>
#> <br>

#plots

 details(
   plot(x = mtcars$mpg, y = mtcars$wt),
   summary = 'Plots')

#> <details closed>
#> <summary> <span title='Click to Expand'> Plots </span> </summary>
#> 
#> ![](https://i.imgur.com/KvHMdeb.png)
#> 
#> </details>
#> <br>

#output options

#character
 details::details('test', output = 'character')

#>  [1] "<details closed>" ""                 "```r"             ""                
#>  [5] "test"             ""                 "```"              ""                
#>  [9] "</details>"       "<br>"            

#clipboard
if(clipr::clipr_available()){

 details::details('test', output = 'clipr')

 clipr::read_clip()

}

#file.edit

 details::details('test', output = 'edit')

#> Warning: Not an interactive enviornment, falling back to output = "console"

#> <details closed>
#> 
#> ```r
#> 
#> test
#> 
#> ```
#> 
#> </details>
#> <br>